renamed saniters pacakage to singular
