+++ sanitizer/tofloat.go
@@ -1,4 +1,4 @@
-package sanitizers
+package sanitizer
 
 import "strconv"
