+++ sanitizer/tofloat_test.go
@@ -1,4 +1,4 @@
-package sanitizers
+package sanitizer
 
 import "testing"
